// src/graphqlschema.d.ts
export type ID = string;
// ...
export interface IUser {
  __typename: "User";
  id: ID;
  username: string;

  /**
   * @deprecated "use id instead"
   */
  sourcegraphID: number;

  /**
   * @deprecated "use emails instead"
   */
  email: string;
  displayName: string | null;
  avatarURL: string | null;
  url: string;
  settingsURL: string;
  createdAt: string;
  updatedAt: string | null;
  siteAdmin: boolean;
  latestSettings: ISettings | null;
  configurationCascade: IConfigurationCascade;
  organizations: IOrgConnection;
  organizationMemberships: IOrganizationMembershipConnection;
  tags: string[];
  activity: IUserActivity;
  emails: IUserEmail[];
  accessTokens: IAccessTokenConnection;
  externalAccounts: IExternalAccountConnection;
  session: ISession;
  viewerCanAdminister: boolean;
  surveyResponses: ISurveyResponse[];
  registryExtensions: IRegistryExtensionConnection;
}
// ...
export interface ISettings {
  __typename: "Settings";
  id: number;
  configuration: IConfiguration;
  subject: ConfigurationSubject;
  author: IUser;
  createdAt: string;

  /**
   * @deprecated "use configuration.contents instead"
   */
  contents: string;
}

export interface IConfiguration {
  __typename: "Configuration";
  contents: string;
  messages: string[];
}

export interface IConfigurationCascade {
  __typename: "ConfigurationCascade";
  subjects: ConfigurationSubject[];
  merged: IConfiguration;
}

export interface IOrgConnection {
  __typename: "OrgConnection";
  nodes: IOrg[];
  totalCount: number;
}
// ...
export interface IRegistryExtensionConnection {
  __typename: "RegistryExtensionConnection";
  nodes: IRegistryExtension[];
  totalCount: number;
  pageInfo: IPageInfo;
  url: string | null;
  error: string | null;
}
// ...
export interface IPageInfo {
  __typename: "PageInfo";
  hasNextPage: boolean;
}

export interface IOrganizationMembershipConnection {
  __typename: "OrganizationMembershipConnection";
  nodes: IOrganizationMembership[];
  totalCount: number;
}

export interface IOrganizationMembership {
  __typename: "OrganizationMembership";
  organization: IOrg;
  user: IUser;
  createdAt: string;
  updatedAt: string;
}

export interface IUserActivity {
  __typename: "UserActivity";
  searchQueries: number;
  pageViews: number;
  codeIntelligenceActions: number;
  lastActiveTime: string | null;
  lastActiveCodeHostIntegrationTime: string | null;
}

export interface IUserEmail {
  __typename: "UserEmail";
  email: string;
  verified: boolean;
  verificationPending: boolean;
  user: IUser;
  viewerCanManuallyVerify: boolean;
}

export interface IAccessTokenConnection {
  __typename: "AccessTokenConnection";
  nodes: IAccessToken[];
  totalCount: number;
  pageInfo: IPageInfo;
}

export interface IAccessToken {
  __typename: "AccessToken";
  id: ID;
  subject: IUser;
  scopes: string[];
  note: string;
  creator: IUser;
  createdAt: string;
  lastUsedAt: string | null;
}

export interface IExternalAccountConnection {
  __typename: "ExternalAccountConnection";
  nodes: IExternalAccount[];
  totalCount: number;
  pageInfo: IPageInfo;
}

export interface IExternalAccount {
  __typename: "ExternalAccount";
  id: ID;
  user: IUser;
  serviceType: string;
  serviceID: string;
  clientID: string;
  accountID: string;
  createdAt: string;
  updatedAt: string;
  refreshURL: string | null;
  accountData: any | null;
}

export interface ISession {
  __typename: "Session";
  canSignOut: boolean;
}

export interface ISurveyResponse {
  __typename: "SurveyResponse";
  id: ID;
  user: IUser | null;
  email: string | null;
  score: number;
  reason: string | null;
  better: string | null;
  createdAt: string;
}
// ...
export interface IDiscussionThreadConnection {
  __typename: "DiscussionThreadConnection";
  nodes: IDiscussionThread[];
  totalCount: number;
  pageInfo: IPageInfo;
}

export interface IDiscussionThread {
  __typename: "DiscussionThread";
  id: ID;
  author: IUser;
  title: string;
  target: DiscussionThreadTarget;
  inlineURL: string | null;
  createdAt: string;
  updatedAt: string;
  archivedAt: string | null;
  comments: IDiscussionCommentConnection;
}

export interface ICommentsOnDiscussionThreadArguments {
  first?: number | null;
}

export type DiscussionThreadTarget = IDiscussionThreadTargetRepo;

export interface IDiscussionThreadTargetRepo {
  __typename: "DiscussionThreadTargetRepo";
  repository: IRepository;
  path: string | null;
  branch: IGitRef | null;
  revision: IGitRef | null;
  selection: IDiscussionThreadTargetRepoSelection | null;
  relativePath: string | null;
  relativeSelection: IDiscussionSelectionRange | null;
}
// ...
export interface IDiscussionCommentConnection {
  __typename: "DiscussionCommentConnection";
  nodes: IDiscussionComment[];
  totalCount: number;
  pageInfo: IPageInfo;
}

export interface IDiscussionComment {
  __typename: "DiscussionComment";
  id: ID;
  thread: IDiscussionThread;
  author: IUser;
  contents: string;
  html: string;
  inlineURL: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface IHtmlOnDiscussionCommentArguments {
  options?: IMarkdownOptions | null;
}

export interface IMarkdownOptions {
  alwaysNil?: string | null;
}
// ...
export interface IDiscussionThreadCreateInput {
  title: string;
  contents: string;
  targetRepo?: IDiscussionThreadTargetRepoInput | null;
}

export interface IDiscussionThreadTargetRepoInput {
  repository: ID;
  path?: string | null;
  branch?: string | null;
  revision?: any | null;
  selection?: IDiscussionThreadTargetRepoSelectionInput | null;
}

export interface IDiscussionThreadTargetRepoSelectionInput {
  startLine: number;
  startCharacter: number;
  endLine: number;
  endCharacter: number;
  linesBefore: string[];
  lines: string[];
  linesAfter: string[];
}